requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-telegram-bot==20.8
openai==1.12.0
python-dotenv==1.0.1
pydantic==2.6.1
instructor==0.6.0
